Applied new Wol.php into old Wol.php
